if文の中身の修正

+++ src/modules/layouts/layouts.repository.ts
@@ -10,7 +10,7 @@ export const layoutsRepository = {
       .eq('user_id', userId)
       .order('created_at', { ascending: false })
       .limit(20);
-    if (data == null || error != null) throw new Error(error.message);
+    if (!data || error) throw new Error(error.message);
     // supabase形式をClassroomLayout[]の型になるように変換
     const formattedLayouts: ClassroomLayout[] = (data || []).map((layout) => ({
       ...layout,
@@ -34,16 +34,21 @@ export const layoutsRepository = {
       })
       .select()
       .single();
-    if (data == null || error != null) throw new Error(error.message);
-    return data;
+    if (!data || error) throw new Error(error.message);
+    // appで使用するClassroomLayout型に変更
+    return {
+      id: data.id,
+      name: data.name,
+      date: data.date,
+      rows: data.rows,
+      cols: data.cols,
+      seats: data.seats as Seat[],
+      students: data.students as Student[],
+    };
   },
   // 座席を削除する
-  async deleteLayout(userId: string, id:string) {
-    const { error } = await supabase
-      .from('layouts')
-      .delete()
-      .eq('user_id', userId)
-      .eq('id', id);
-    if (error != null) throw new Error(error.message);
+  async deleteLayout(userId: string, id: string) {
+    const { error } = await supabase.from('layouts').delete().eq('user_id', userId).eq('id', id);
+    if (error) throw new Error(error.message);
   },
 };
